% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_licorfile.R
\name{analyse_licorfile}
\alias{analyse_licorfile}
\title{A function that analyses a licorfile for visual examination}
\usage{
analyse_licorfile(
  pathname,
  parameters = c("elapsed", "A", "Ci", "Qin", "TleafCnd", "VPDleaf", "RHcham", "CO2_r",
    "gsw"),
  x_axis = c("elapsed"),
  exclude = c("CO2_r"),
  calculate_means = FALSE,
  plot_individuals = TRUE
)
}
\arguments{
\item{pathname}{The licor file that will be analysed.}

\item{parameters}{A vector of the recuired parameters}

\item{x_axis}{A vector with the string of the x_axis name}

\item{exclude}{A vector of parameters to exclude from the normalized plot}

\item{calculate_means}{A boolean if the means should be calculated and returned}

\item{plot_individuals}{A boolean that regulates if all the individuals are plotted}
}
\value{
Returns a list of means if calculate means is true
}
\description{
A function that analyses a licorfile for visual examination.
  Gives a quick overview of the data.
}
\examples{

}
\author{
Sam Loontjens
}
