% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licorfile_6800.R
\name{read_licorfile_6800}
\alias{read_licorfile_6800}
\title{A function that reads a 6800 licorfile.}
\usage{
read_licorfile_6800(
  filepath,
  sheetnumber = 1,
  parameters = c(),
  numeric = TRUE
)
}
\arguments{
\item{filepath}{A string with the filepath to read.}

\item{sheetnumber}{An integer which sheet to read from the file.}

\item{parameters}{A list of desired parameters to extract. Default is everything.}

\item{change_names}{A boolean that regulates if the names from the Li-6400 files are changed
to the Li-6800 format. Default is TRUE.}
}
\value{
Returns a dataframe with only the desired parameters.
}
\description{
A function that reads a licorfile.
  It returns a clean dataframe with only the desired parametes.

  No dots in the file path accepted, except one for the file type.
  The functions of this package expect the parameter names of a Li-6800, so
  change_names = TRUE is recommended.
  Li-6800 manual page 226
  Li-6400 manual page 21 (81 of 1324)
}
\examples{
parameters_to_extract <- c("A, Ci, elapsed, Qin")
dataframe <- read_licorfile("20210226 PI 50-100 75RH 400CO2 T.xlsx",
                         parameters_to_extract)

}
\author{
Sam Loontjens
}
