% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_parameter.R
\name{round_parameter}
\alias{round_parameter}
\title{A function that rounds one column or vector.}
\usage{
round_parameter(column)
}
\arguments{
\item{column}{A column or vector of a dataframe}
}
\value{
Returns a new column or vector that is rounded.
}
\description{
A function that rounds one column or vector.
  It returns a new column or vector that is rounded.
}
\examples{
mydata$Qin <- round_parameter(mydata$Qin)

}
